+++ include/uapi/linux/provenance.h
@@ -31,7 +31,7 @@
 	"."xstr (CAMFLOW_VERSION_MINOR)					\
 	"."xstr (CAMFLOW_VERSION_PATCH)					\
 
-#define CAMFLOW_COMMIT "071e407a9556ac0180374a7e047c0d1fa342926e"
+#define CAMFLOW_COMMIT "e84608debb699332c381b37de7c3f65ef0500b6e"
 
 #define PROVENANCE_HASH                 "sha256"